Fix/betaissues (#675)
- Reorganized PrefKey list to be less confusing.
- Disengage custom shortcut keyboard after 100 key repeat to prevent possibly endless loop if keyUp event never arrives due to any circumstance.
- Don't relinquish control over brightness keys with no external display connected if fine brightness OSD scale is active
- Register DDC command touched status. When write on startup enabled, apply only touched command values.
- Fixed custom key shortcuts going runaway when menu was opened during a keyrepeat streak.
