+++ MonitorControl/Enums/PrefKey.swift
@@ -1,65 +1,13 @@
 //  Copyright © MonitorControl. @JoniVR, @theOneyouseek, @waydabber and others
 
 enum PrefKey: String {
-  // Enable mute DDC for display
-  case enableMuteUnmute
+  /* -- App-wide settings -- */
 
   // Sparkle automatic checks
   case SUEnableAutomaticChecks
 
   // Receive beta updates?
   case isBetaChannel // This is not added to Preferences yet as it will be needed in the future only.
-
-  // Hide OSD for display
-  case hideOsd
-
-  // Longer delay DDC for display
-  case longerDelay
-
-  // DDC polling mode for display
-  case pollingMode
-
-  // DDC polling count for display
-  case pollingCount
-
-  // Display should avoid gamma table manipulation and use shades instead (to coexist with other apps doing gamma manipulation)
-  case avoidGamma
-
-  // Command value display
-  case value
-
-  // Min command value display
-  case minDDCOverride
-
-  // Max command value display
-  case maxDDC
-
-  // Max user override command value display
-  case maxDDCOverride
-
-  // Max command value display
-  case curveDDC
-
-  // Is the specific control is set as unavailable for display?
-  case unavailableDDC
-
-  // Invert DDC scale?
-  case invertDDC
-
-  // Override DDC control command code
-  case remapDDC
-
-  // User assigned audio device name for display
-  case audioDeviceNameOverride
-
-  // Display disabled for keyboard control
-  case isDisabled
-
-  // Force software mode for display
-  case forceSw
-
-  // Software brightness for display
-  case SwBrightness
 
   // Build number
   case buildNumber
@@ -94,9 +42,6 @@ enum PrefKey: String {
   // Lower via software after brightness
   case disableCombinedBrightness
 
-  // Lower via software after brightness
-  case combinedBrightnessSwitchingPoint
-
   // Use separated OSD scale for combined brightness
   case separateCombinedScale
 
@@ -114,9 +59,6 @@ enum PrefKey: String {
 
   // Show tick marks for sliders
   case showTickMarks
-
-  // Friendly name changed
-  case friendlyName
 
   // Instead of assuming default values, enable read or write upon startup (according to readDDCInsteadOfRestoreValues)
   case enableDDCDuringStartup
@@ -159,6 +101,73 @@ enum PrefKey: String {
 
   // Combine sliders for all displays
   case slidersCombine
+
+  /* -- Display specific settings */
+
+  // Enable mute DDC for display
+  case enableMuteUnmute
+
+  // Hide OSD for display
+  case hideOsd
+
+  // Longer delay DDC for display
+  case longerDelay
+
+  // DDC polling mode for display
+  case pollingMode
+
+  // DDC polling count for display
+  case pollingCount
+
+  // Display should avoid gamma table manipulation and use shades instead (to coexist with other apps doing gamma manipulation)
+  case avoidGamma
+
+  // User assigned audio device name for display
+  case audioDeviceNameOverride
+
+  // Display disabled for keyboard control
+  case isDisabled
+
+  // Force software mode for display
+  case forceSw
+
+  // Software brightness for display
+  case SwBrightness
+
+  // Combined brightness switching point
+  case combinedBrightnessSwitchingPoint
+
+  // Friendly name
+  case friendlyName
+
+  /* -- Display+Command specific settings -- */
+
+  // Command value display
+  case value
+
+  // Was the setting ever changed by the user?
+  case isTouched
+
+  // Min command value display
+  case minDDCOverride
+
+  // Max command value display
+  case maxDDC
+
+  // Max user override command value display
+  case maxDDCOverride
+
+  // Max command value display
+  case curveDDC
+
+  // Is the specific control is set as unavailable for display?
+  case unavailableDDC
+
+  // Invert DDC scale?
+  case invertDDC
+
+  // Override DDC control command code
+  case remapDDC
 }
 
 enum PollingMode: Int {
