+++ MonitorControl/Model/OtherDisplay.swift
@@ -1,6 +1,5 @@
 //  Copyright © MonitorControl. @JoniVR, @theOneyouseek, @waydabber and others
 
-import AVFoundation
 import Cocoa
 import IOKit
 import os.log
@@ -13,7 +12,6 @@ class OtherDisplay: Display {
   var arm64avService: IOAVService?
   var isDiscouraged: Bool = false
   let DDC_MAX_DETECT_LIMIT: Int = 100
-  private var audioPlayer: AVAudioPlayer?
   var pollingCount: Int {
     get {
       switch self.readPrefAsInt(key: .pollingMode) {
@@ -114,7 +112,8 @@ class OtherDisplay: Display {
       os_log("- Minimum DDC value: %{public}@ (overrides 0)", type: .info, String(self.readPrefAsInt(key: .minDDCOverride, for: command)))
       os_log("- Maximum DDC value: %{public}@ (overrides %{public}@)", type: .info, String(self.readPrefAsInt(key: .maxDDC, for: command)), String(maxDDCValue))
       os_log("- Current internal value: %{public}@", type: .info, String(self.readPrefAsFloat(for: command)))
-      if prefs.bool(forKey: PrefKey.enableDDCDuringStartup.rawValue), !prefs.bool(forKey: PrefKey.readDDCInsteadOfRestoreValues.rawValue), !app.safeMode {
+      os_log("- Command status: %{public}@", type: .info, self.readPrefAsBool(key: .isTouched, for: command) ? "Touched" : "Untouched")
+      if self.readPrefAsBool(key: .isTouched, for: command), prefs.bool(forKey: PrefKey.enableDDCDuringStartup.rawValue), !prefs.bool(forKey: PrefKey.readDDCInsteadOfRestoreValues.rawValue), !app.safeMode {
         os_log("- Writing last saved DDC values.", type: .info, self.name, String(reflecting: command))
         _ = self.writeDDCValues(command: command, value: currentDDCValue)
       }
@@ -390,6 +389,7 @@ class OtherDisplay: Display {
         } else {
           success = self.ddc?.write(command: command.rawValue, value: value, errorRecoveryWaitTime: 2000) ?? false
         }
+        self.savePref(true, key: PrefKey.isTouched, for: command) // We deliberatly consider the value tuched no matter if the call succeeded
       }
     }
     return success
@@ -484,19 +484,6 @@ class OtherDisplay: Display {
       value = 1 - value
     }
     return max(min(value, 1), 0)
-  }
-
-  func playVolumeChangedSound() {
-    guard let preferences = app.getSystemPreferences(), let hasSoundEnabled = preferences["com.apple.sound.beep.feedback"] as? Int, hasSoundEnabled == 1 else {
-      return
-    }
-    do {
-      self.audioPlayer = try AVAudioPlayer(contentsOf: URL(fileURLWithPath: "/System/Library/LoginPlugins/BezelServices.loginPlugin/Contents/Resources/volume.aiff"))
-      self.audioPlayer?.volume = 1
-      self.audioPlayer?.play()
-    } catch {
-      os_log("%{public}@", type: .error, error.localizedDescription)
-    }
   }
 
   func combinedBrightnessSwitchingValue() -> Float {
